correct download url set to firestore
